DEMRUM-6028: Ignore frames delivered while inactive
* Add an actor-owned isActive flag to SlowFrameLogic, gated in
  handleFrame and toggled by the lifecycle handlers
* Closes a race between the frame-consuming AsyncStream task and the
  lifecycle-notification task: a frame already in flight when the app
  resigns active could previously re-arm the heartbeat after
  appWillResignActive had reset it, letting the watchdog count a
  spurious frozen frame while backgrounded
